requirements.txt
flask
streamlit
requests
pypdf
faiss-cpu
langchain
vertexai
numpy
cryptography>=3.1
